backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
opencv-python==4.7.0.72
mediapipe==0.10.9
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
scipy==1.10.1
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0